fix: execute the installed Wordcell command
