+++ .gitignore
@@ -3,3 +3,5 @@ node_modules/
 next-env.d.ts
 *.tsbuildinfo
 .DS_Store
+.vercel
+.env*
