+++ src/kb-alias.ts
@@ -4,7 +4,7 @@
  * `wordcell`; this entry point keeps existing `kb …` invocations working for
  * one minor cycle and is removed in 0.21.0.
  */
-import { runExecutable } from "./cli.js";
+import { runExecutable } from "./cli-program.js";
 
 export const KB_ALIAS_NOTICE = "kb is now wordcell; the kb alias is removed in 0.21.0";
